chore: move ecrecover implementation to aurora-engine-sdk
