+++ engine-sdk/Cargo.toml
@@ -13,14 +13,18 @@ autobenches = false
 [dependencies]
 aurora-engine-types.workspace = true
 base64.workspace = true
+libsecp256k1 = { workspace = true, features = ["static-context", "hmac"] }
 sha2.workspace = true
 sha3.workspace = true
+
+[dev-dependencies]
+hex.workspace = true
 
 [lints]
 workspace = true
 
 [features]
-std = ["aurora-engine-types/std", "sha3/std", "sha2/std", "base64/std"]
+std = ["aurora-engine-types/std", "libsecp256k1/std", "sha3/std", "sha2/std", "base64/std"]
 contract = []
 log = []
 all-promise-actions = []
