+++ engine-sdk/src/lib.rs
@@ -2,9 +2,9 @@
 // All `as` conversions in this code base have been carefully reviewed and are safe.
 #![allow(clippy::as_conversions)]
 
+use crate::prelude::{Address, H256, STORAGE_PRICE_PER_BYTE};
 #[cfg(feature = "contract")]
-use crate::prelude::{Address, Vec, U256};
-use crate::prelude::{H256, STORAGE_PRICE_PER_BYTE};
+use crate::prelude::{Vec, U256};
 pub use types::keccak;
 
 pub mod base64;
@@ -181,6 +181,29 @@ pub fn ecrecover(hash: H256, signature: &[u8]) -> Result<Address, ECRecoverErr> 
     }
 }
 
+#[cfg(not(feature = "contract"))]
+pub fn ecrecover(hash: H256, signature: &[u8]) -> Result<Address, ECRecoverErr> {
+    use sha3::Digest;
+
+    let hash = libsecp256k1::Message::parse_slice(hash.as_bytes()).map_err(|_| ECRecoverErr)?;
+    let v = signature[64];
+    let signature = libsecp256k1::Signature::parse_standard_slice(&signature[0..64])
+        .map_err(|_| ECRecoverErr)?;
+    let bit = match v {
+        0..=26 => v,
+        _ => v - 27,
+    };
+    let recovery_id = libsecp256k1::RecoveryId::parse(bit).map_err(|_| ECRecoverErr)?;
+
+    libsecp256k1::recover(&hash, &signature, &recovery_id)
+        .map_err(|_| ECRecoverErr)
+        .and_then(|public_key| {
+            // recover returns a 65-byte key, but addresses come from the raw 64-byte key
+            let r = sha3::Keccak256::digest(&public_key.serialize()[1..]);
+            Address::try_from_slice(&r[12..]).map_err(|_| ECRecoverErr)
+        })
+}
+
 #[cfg(feature = "contract")]
 pub fn log(data: &str) {
     log_utf8(data.as_bytes());
@@ -217,5 +240,34 @@ impl ECRecoverErr {
 impl AsRef<[u8]> for ECRecoverErr {
     fn as_ref(&self) -> &[u8] {
         self.as_str().as_bytes()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use aurora_engine_types::types::Address;
+    use aurora_engine_types::H256;
+
+    const SIGNATURE_LENGTH: usize = 65;
+
+    fn ecverify(hash: H256, signature: &[u8], signer: Address) -> bool {
+        matches!(ecrecover(hash, signature[0..SIGNATURE_LENGTH].try_into().unwrap()), Ok(s) if s == signer)
+    }
+
+    #[test]
+    fn test_ecverify() {
+        let hash = H256::from_slice(
+            &hex::decode("1111111111111111111111111111111111111111111111111111111111111111")
+                .unwrap(),
+        );
+        let signature =
+            &hex::decode("b9f0bb08640d3c1c00761cdd0121209268f6fd3816bc98b9e6f3cc77bf82b69812ac7a61788a0fdc0e19180f14c945a8e1088a27d92a74dce81c0981fb6447441b")
+                .unwrap();
+        let signer = Address::try_from_slice(
+            &hex::decode("1563915e194D8CfBA1943570603F7606A3115508").unwrap(),
+        )
+        .unwrap();
+        assert!(ecverify(hash, signature, signer));
     }
 }
